+++ src/lib/gear/types.ts
@@ -10,6 +10,9 @@ export const GearSetupTypes = ['melee', 'range', 'mage', 'misc', 'skilling', 'wi
 
 export type GearSetupType = (typeof GearSetupTypes)[number];
 
+export const PrimaryGearSetupTypes = ['melee', 'range', 'mage'] as const;
+export type PrimaryGearSetupType = (typeof PrimaryGearSetupTypes)[number];
+
 export enum GearStat {
 	AttackStab = 'attack_stab',
 	AttackSlash = 'attack_slash',
