+++ src/lib/minions/types.ts
@@ -6,7 +6,7 @@ import type { Item } from 'oldschooljs/dist/meta/types';
 import type SimpleMonster from 'oldschooljs/dist/structures/SimpleMonster';
 
 import type { ClueTier } from '../clues/clueTiers';
-import type { BitField, PerkTier } from '../constants';
+import type { BitField, PerkTier, PvMMethod } from '../constants';
 import type { GearSetupType, GearStat, OffenceGearStat } from '../gear/types';
 import type { POHBoosts } from '../poh';
 import type { MinigameName } from '../settings/minigames';
@@ -177,7 +177,8 @@ export interface AddMonsterXpParams {
 
 export interface ResolveAttackStylesParams {
 	monsterID: number | undefined;
-	boostMethod?: string[];
+	boostMethod?: PvMMethod[];
+	attackStyles: AttackStyles[];
 }
 
 export interface BlowpipeData {
